% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_badge.R
\name{add_badge_status}
\alias{add_badge_status}
\alias{add_badge_codefactor}
\alias{add_badge_zenodo}
\title{Add badges to README}
\usage{
add_badge_status(
  status = c("concept", "wip", "suspended", "abandoned", "active", "inactive",
    "unsupported", "moved"),
  path = NULL,
  .url = NULL
)

add_badge_codefactor(repo = NULL, path = NULL)

add_badge_zenodo(repo = NULL, path = NULL)
}
\arguments{
\item{status}{A character value for status to be assigned to project. This
can be either \emph{"concept"}, \emph{"wip"}, \emph{"suspended"}, \emph{"abandoned"},
\emph{"active"}, \emph{"inactive"}, \emph{"unsupported"}, or \emph{"moved"}.}

\item{path}{Path to file to add repostatus badge to. Set to NULL by default
which would indicate that a README file in the root directory of the
project is the target file.}

\item{.url}{If \code{status} is "moved", the URL to which the repository has moved
to. Otherwise NULL.}

\item{repo}{Short remote git repository name. If NULL, is determined based
on current git settings.}
}
\value{
An entry to the badge section of the README of the repository.
Otherwise, a print of the markdown text for the status badge.
}
\description{
Add badges to README
}
\examples{
if (FALSE) {
  add_badge_status("wip")
  add_badge_codefactor(repo = "katilingban/pakete")
  add_badge_zenodo(repo = "katilingban/pakete")
}

}
