% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_badge.R
\name{add_badge_codefactor}
\alias{add_badge_codefactor}
\title{Add CodeFactor badget}
\usage{
add_badge_codefactor(repo = NULL, path = NULL)
}
\arguments{
\item{repo}{Short remote git repository name. If NULL, is determined based
on current git settings.}

\item{path}{Path to file to add repostatus badge to. Set to NULL by default
which would indicate that a README file in the root directory of the
project is the target file.}
}
\value{
An entry to the badge section of the README of the repository.
  Otherwise, a print of the markdown text for the status badge.
}
\description{
Assumes that repository is already activated in CodeFactor.
}
\examples{
if (FALSE) {add_badge_codefactor(repo = "katilingban/pakete")}

}
