% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_badge.R
\name{add_badge_status}
\alias{add_badge_status}
\title{Add repostatus badge}
\usage{
add_badge_status(
  status = c("concept", "wip", "suspended", "abandoned", "active", "inactive",
    "unsupported", "moved"),
  path = NULL,
  .url = NULL
)
}
\arguments{
\item{status}{A character value for status to be assigned to project. This
can be either *"concept"*, *"wip"*, *"suspended"*, *"abandoned"*,
*"active"*, *"inactive"*, *"unsupported"*, or *"moved"*.}

\item{path}{Path to file to add repostatus badge to. Set to NULL by default
which would indicate that a README file in the root directory of the
project is the target file.}

\item{.url}{If `status` is "moved", the URL to which the repository has moved
to. Otherwise NULL.}
}
\value{
An entry to the badge section of the README of the repository.
  Otherwise, a print of the markdown text for the status badge.
}
\description{
Add repostatus badge
}
\examples{
add_badge_status(
  "wip",
  path = system.file("examples", "README.Rmd", package = "pakete")
)

}
