% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_codecov.R
\name{add_codecov}
\alias{add_codecov}
\title{Add Codecov token snippet}
\usage{
add_codecov()
}
\value{
An entry in the \code{codecov.yml} file for Codecov token specifications
suitable for GitHub Actions.
}
\description{
Add Codecov token snippet
}
\examples{
if (interactive()) add_codecov()

}
