% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_contributing.R
\name{add_contributing}
\alias{add_contributing}
\title{Add contributing markdown}
\usage{
add_contributing(repo = NULL, overwrite = FALSE)
}
\arguments{
\item{repo}{Short remote git repository name. If NULL, is determined based
on current git settings.}

\item{overwrite}{Logical. Should an existing CONTRIBUTING.md file be
overwritten? Default to FALSE.}
}
\value{
A CONTRIBUTING.md file in the \code{.github} directory
}
\description{
Add contributing markdown
}
\examples{
if (FALSE) add_contributing("katilingban/pakete")

}
