% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_github_action.R
\name{add_gha_workflow}
\alias{add_gha_workflow}
\title{Add GHA workflow}
\usage{
add_gha_workflow(gha_name, gha_file_name, gha_file_path, overwrite)
}
\description{
Add GHA workflow
}
\keyword{internal}
