% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_github_action.R
\name{add_github_action}
\alias{add_github_action}
\title{Add GitHub Actions}
\usage{
add_github_action(gha_name = NULL, repo = NULL, overwrite = FALSE)
}
\arguments{
\item{gha_name}{Short name of GitHub Action to add. Currently supports
GitHub Action for Netlify deployment of testing version of pkgdown website
(\emph{"netlify"}) and GitHub Action for mirroring to Codeberg
(\emph{"mirror-codeberg"}).}

\item{repo}{Short remote git repository name. If NULL, is determined based
on current git settings. Evaluated only if \code{gha_name} is
\emph{"mirror-codeberg"}.}

\item{overwrite}{Logical. Should an existing GitHub Action be overwritten?
Default is FALSE.}
}
\value{
A GitHub Action YAML file of the specified workflow to be added in
\code{.github/workflows} directory.
}
\description{
Add GitHub Actions
}
\examples{
if (FALSE) add_github_action(gha_name = "netlify")

}
