% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_issue_templates.R
\name{add_issue_initial_release}
\alias{add_issue_initial_release}
\title{Add issue templates}
\usage{
add_issue_initial_release(path = ".github/ISSUE_TEMPLATE")
}
\arguments{
\item{path}{Path to file to add issue template into. Set to
".github/ISSUE_TEMPLATE" which is the default location specified by
GitHub.}
}
\value{
A specified issue template markdown file in the specified `path`.
}
\description{
Add issue templates
}
\examples{
if (interactive()) add_issue_initial_release()

}
