% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_issue_templates.R
\name{add_issue_template_}
\alias{add_issue_template_}
\title{Add issue template}
\usage{
add_issue_template_(issue, template_file_name, template_file_path, overwrite)
}
\description{
Add issue template
}
\keyword{internal}
