% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_contributing.R
\name{create_contributing}
\alias{create_contributing}
\title{Create CONTRIBUTING.md}
\usage{
create_contributing(repo)
}
\arguments{
\item{repo}{Short remote git repository name. If NULL, is determined based
on current git settings.}
}
\value{
A CONTRIBUTING text.
}
\description{
Create CONTRIBUTING.md
}
\examples{
if (FALSE) create_contributing("katilingban/pakete")

}
\keyword{internal}
