% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_github_repository}
\alias{get_github_repository}
\title{Get current remote GitHub repository}
\usage{
get_github_repository(full = FALSE)
}
\arguments{
\item{full}{Logical. Should full GitHub remote URL be extracted? Default to
FALSE.}
}
\value{
GitHub remote repository.
}
\description{
Get current remote GitHub repository
}
\examples{
if (FALSE) get_github_repository()

}
\keyword{internal}
