% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_github_repository_id}
\alias{get_github_repository_id}
\title{Get GitHub repository ID}
\usage{
get_github_repository_id(repo)
}
\arguments{
\item{repo}{Short remote git repository name for current project.}
}
\value{
An integer value for the GitHub repository ID.
}
\description{
Get GitHub repository ID
}
\examples{
if (FALSE) get_gihtub_repository_id(repo = "katilingban/pakete")

}
\keyword{internal}
