% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_readme_path}
\alias{get_readme_path}
\title{Get file path to README file}
\usage{
get_readme_path(path = NULL)
}
\arguments{
\item{path}{Path to README file. Set to NULL by default which would indicate
that a README file is in the root directory of the project.}
}
\value{
Path to README file.
}
\description{
Get file path to README file
}
\examples{
if (FALSE) get_readme_path()

}
\keyword{internal}
