% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{insert_readme_badge_text}
\alias{insert_readme_badge_text}
\title{Insert text to badge section of README lines}
\usage{
insert_readme_badge_text(badge_text = NULL, readme_lines = NULL, path)
}
\arguments{
\item{badge_text}{A character value for text to insert into badge section of
README lines.}

\item{readme_lines}{A character vector of lines of text in README.}

\item{path}{Path to file to add repostatus badge to.}
}
\value{
A character vector of README lines of text.
}
\description{
Insert text to badge section of README lines
}
\examples{
if (FALSE) get_readme_badge_lines(readme_lines)

}
\keyword{internal}
