% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_contributing.R
\name{save_contributing}
\alias{save_contributing}
\title{Save CONTRIBUTING.md}
\usage{
save_contributing(contrib_lines)
}
\arguments{
\item{contrib_lines}{CONTRIBUTING text to save into CONTRIBUTING.md.}
}
\value{
A CONTRIBUTING.md file in the \code{.github} directory
}
\description{
Save CONTRIBUTING.md
}
\examples{
if (FALSE) save_contributing(contrib_lines)

}
\keyword{internal}
