% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_github_action.R
\name{save_gha_workflow_codeberg_mirror}
\alias{save_gha_workflow_codeberg_mirror}
\title{Save Codeberg mirror GitHub Actions workflow}
\usage{
save_gha_workflow_codeberg_mirror(gha_text)
}
\arguments{
\item{gha_text}{Codeberg mirror GitHub Actions workflow text to save into
mirror-codeberg.yaml.}
}
\value{
A mirror-codeberg.yaml file in the \code{.github/workflows} directory
}
\description{
Save Codeberg mirror GitHub Actions workflow
}
\examples{
if (FALSE) save_gha_workflow_codeberg_mirror(gha_text)

}
\keyword{internal}
