% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_contributing.R
\name{add_contributing}
\alias{add_contributing}
\title{Add contributing document}
\usage{
add_contributing(repo = NULL)
}
\arguments{
\item{repo}{Short remote git repository name. If NULL, is determined based
on current git settings.}
}
\value{
A CONTRIBUTING.md file in the `.github` directory
}
\description{
Add contributing document
}
\examples{
if (interactive()) add_contributing("katilingban/pakete")

}
