% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pakete-package.R
\docType{package}
\name{pakete}
\alias{pakete-package}
\alias{pakete}
\title{Tools and utilities for package development currently not available from
usual development tools. These are mostly linked to personal preferences
during the development process. They assist in making routine and repetitive
tasks easily implementable.}
\description{
Tools and utilities for package development currently not available from usual development tools. These are mostly linked to personal preferences during the development process. They assist in making routine and repetitive tasks easily implementable.
}
\author{
\strong{Maintainer}: Ernest Guevarra \email{ernest@guevarra.io} (\href{https://orcid.org/0000-0002-4887-4415}{ORCID})

}
\keyword{internal}
